+++ src/main.jsx
@@ -4,16 +4,23 @@ import { BrowserRouter } from 'react-router-dom'
 import App from './App'
 import { AuthProvider } from './contexts/AuthContext'
 import { NotificationProvider } from './contexts/NotificationContext'
+import { SentryErrorBoundary } from './components/common/SentryErrorBoundary'
+import { initSentry } from './config/sentry'
 import './index.css'
+
+// Initialize Sentry before rendering
+initSentry()
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
-    <BrowserRouter>
-      <AuthProvider>
-        <NotificationProvider>
-          <App />
-        </NotificationProvider>
-      </AuthProvider>
-    </BrowserRouter>
+    <SentryErrorBoundary>
+      <BrowserRouter>
+        <AuthProvider>
+          <NotificationProvider>
+            <App />
+          </NotificationProvider>
+        </AuthProvider>
+      </BrowserRouter>
+    </SentryErrorBoundary>
   </React.StrictMode>
 )